Move hello-world component to module directory
Intend to group components in related blocks, not listed in a single directory
